App: Extract BackupPolicy into its own file
No code changes except to add AppExport to BackupPolicy
